The beginnings of logic implementation.
- Removed OverwatchProfileParser and OverwatchWebClient.
I realised that these would only be used by OverwatchClient and it would make more sense to just include their functionality in OverwatchClient.

- Added some logic to Overwatch Client.
No actual stat retrieval yet, but it's very close.

- Added OverwatchSearchResult.
This class is an internal class that'll be used to deserialize results from the non-public player search API. It will never be seen by the end user.