Reworked player level retrieval
We're no longer using the non-public API to fetch the player level along with aliases.
Instead, we're parsing the level and prestige data from the webpage and leaving out aliases altogether.
Aliases can be fetched separately with a new method: OverwatchPlayer.GetPlayerAliasesAsync. The non-public API will only be used to get aliases now.

In addition, this commit has a few minor tweaks to a few methods and the readme was updated with badges and examples.